# Refactor sorted map encode to use fewer buffers for nested maps.